create the single-post view

+++ controllers/api/comment-routes.js
@@ -28,7 +28,8 @@ router.post('/', (req, res) => {
 });
 
 router.delete('/:id', (req, res) => {
-    Comment.destroy({
+    if(req.session){
+        Comment.destroy({
         where: {
             id: req.params.id
         }
@@ -44,6 +45,7 @@ router.delete('/:id', (req, res) => {
             console.log(err);
             res.status(500).json(err);
         });
+    }
 });
 
 module.exports = router;
